api/requirements.txt
celery
Flask
redis
rollbar
blinker
python-dotenv
requests
